Micah debugging the add job field
I had to fix one of the relations in the models index to make the clients show up properly when added to the post. i fixed minor things in the public javacript file. The complete var was named to isDone to be compatiable with the associated variable in the jobs table. it doesn't work if it isn't named the same. i also changed the .value to .checked so that way it returns true or false. fixed an old seed testing mistake i made. and in the handlebars i set the value as the id of the assocated field and what actual shows as the name. IE. clients value is the client id, but you can select the client name. all this should with the database, waiting on the post request to finish